move Table refs to TableInstance
